questlab/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Werkzeug==3.1.3
Pillow==12.0.0
python-dotenv==1.1.1
pytest==7.4.0
pytest-flask==1.2.0
WTForms==3.2.1
python-magic==0.4.27  # Optional - only if you use the advanced validation
Flask-Login==0.6.3